+++ Docs/Whats_New.md
@@ -16,10 +16,10 @@ For best results, use a Markdown reader.*
 MoltenVK 1.0.31
 ---------------
 
-Released 2018/01/12
+Released 2018/01/15
 
 - Support runtime config via runtime environment variables
-- Add full texture swizzling to config, and disable it by default.
+- Add full ImageView swizzling to config, and disable it by default.
 - Add GPU switching to config, and enable it by default.
 - Add queue family specialization to config, and disable it by default.
 - Enable synchronous queue submits as config default.